backend/ai-service/requirements.txt
# Senior MHealth AI Service - Dependencies
# Text analysis using Gemini API + Speech-to-Text

# Core dependencies
google-cloud-aiplatform>=1.38.0  # Vertex AI SDK for Gemini API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # includes uvloop for performance
pydantic>=2.0.0
python-dotenv>=1.0.0

# Cloud integration
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-cloud-speech>=2.21.0  # Speech-to-Text API

# Utilities
httpx>=0.25.0
python-multipart>=0.0.6
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
